Fix parsing of IPv4 address in SOCKS4A
According to the protocol a domain is used instead of an IPv4 address if
the address is 0.0.0.x with x non-zero.

Since (0x00000000 & 0x000000ff) == 0x00000000 the IP address 0.0.0.0 was
considered to be a domain as well, so we need to check if not only the
first 3 octets are zero but if the last octet is non-zero as well.

I'm not sure if any client sets 0.0.0.0 as IP address. I didn't see a
real life issue with this, but nonetheless let's try to be true to the
specification.