Add support for connection filtering
This is not actually doing anything yet, since there is never a
connection filter or factory instantiated. One has to implement a custom
connection filter and a factory which instantiates objects of this
filter.

I've chosen this approach, since it allows to have more complex
filtering depending on your need / use case.
Instances of connection filters are created per connection, since it
allows to have a per connection state for each filter, which then could
do some caching or cleanup or whatever is necessary.